+++ frontend/src/metabase-types/api/collection.ts
@@ -112,7 +112,6 @@ export interface CollectionItem {
   setCollection?: (collection: Pick<Collection, "id">) => void;
   setCollectionPreview?: (isEnabled: boolean) => void;
   update?: (update: Partial<CollectionItem>) => Promise<any>;
-  delete?: () => Promise<void>;
 }
 
 export interface CollectionListQuery {
