+++ frontend/src/metabase/collections/components/BulkActions/index.tsx
@@ -1,0 +1,1 @@
+export * from "./BulkActions";
