+++ frontend/src/metabase/components/Modal/Modal.tsx
@@ -3,13 +3,12 @@ import { RoutelessFullPageModal } from "metabase/components/Modal/RoutelessFullP
 import type { WindowModalProps } from "metabase/components/Modal/WindowModal";
 import { WindowModal } from "metabase/components/Modal/WindowModal";
 
-const Modal = ({
-  full = false,
-  ...props
-}: {
+export type ModalProps = {
   full?: boolean;
   isOpen?: boolean;
-} & (WindowModalProps & FullPageModalProps)) => {
+} & (WindowModalProps & FullPageModalProps);
+
+const Modal = ({ full = false, ...props }: ModalProps) => {
   if (full) {
     return props.isOpen ? <RoutelessFullPageModal {...props} /> : null;
   } else {
