+++ frontend/src/metabase/dashboard/components/DashboardInfoSidebar/DashboardInfoSidebar.tsx
@@ -43,9 +43,6 @@ export function DashboardInfoSidebar({
   const currentUser = useSelector(getUser);
   const dispatch = useDispatch();
 
-  const showCaching =
-    PLUGIN_CACHING.isEnabled() && MetabaseSettings.get("enable-query-caching");
-
   const handleDescriptionChange = useCallback(
     (description: string) => {
       setDashboardAttribute("description", description);
@@ -70,7 +67,13 @@ export function DashboardInfoSidebar({
   );
 
   const autoApplyFilterToggleId = useUniqueId();
+
   const canWrite = dashboard.can_write && !dashboard.archived;
+
+  const showCaching =
+    canWrite &&
+    PLUGIN_CACHING.isEnabled() &&
+    MetabaseSettings.get("enable-query-caching");
 
   return (
     <DashboardInfoSidebarRoot data-testid="sidebar-right">
