+++ frontend/src/metabase/entities/questions.js
@@ -2,7 +2,6 @@ import { updateIn } from "icepick";
 import { t } from "ttag";
 
 import { cardApi } from "metabase/api";
-import { undoSetArchived } from "metabase/archive/utils";
 import { canonicalCollectionId } from "metabase/collections/utils";
 import Collections, {
   getCollectionType,
@@ -63,7 +62,7 @@ const Questions = createEntity({
       Questions.actions.update(
         { id: card.id },
         { archived },
-        undoSetArchived(card.name, archived, opts),
+        undo(opts, getLabel(card), archived ? t`trashed` : t`restored`),
       ),
 
     setCollection: (card, collection, opts) => {
